shell: prepare to publicize flux/shell.h
Prepare to make a shell plugin interface public as <flux/shell.h>.

Move struct flux_shell definition into an "internal.h" header and
keep existing public flux_shell_* functions in shell.h.

Add shell.h to src/include/flux so that existing components may
include it as <flux/shell.h>.

Update current users. No interfaces have been added or removed.
