+++ src/include/flux/shell.h
@@ -1,5 +1,5 @@
 /************************************************************\
- * Copyright 2019 Lawrence Livermore National Security, LLC
+ * Copyright 2014 Lawrence Livermore National Security, LLC
  * (c.f. AUTHORS, NOTICE.LLNS, COPYING)
  *
  * This file is part of the Flux resource manager framework.
@@ -8,17 +8,7 @@
  * SPDX-License-Identifier: LGPL-3.0
 \************************************************************/
 
-#ifndef SHELL_KILL_H
-#define SHELL_KILL_H
-
-#include <flux/shell.h>
-
-/*  Initialize shell kill event handler
+/* Allow in-tree programs to #include <flux/shell.h> like out-of-tree would.
  */
-int kill_event_init (flux_shell_t *shell);
 
-#endif /* !SHELL_KILL_H */
-
-/*
- * vi:tabstop=4 shiftwidth=4 expandtab
- */
+#include "src/shell/shell.h"
