+++ src/shell/svc.c
@@ -33,9 +33,9 @@
 #include <string.h>
 #include <flux/core.h>
 
+#include "internal.h"
 #include "task.h"
 #include "io.h"
-#include "shell.h"
 #include "svc.h"
 
 #define TOPIC_STRING_SIZE  128
